fix: use marked first
